refactor: add class decorator to create singleton class (#69)
