+++ src/generate/templates/component.ts
